# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KYC Service API
  description: |
    Production-grade KYC verification microservice — authentication, profile submission,
    risk scoring, and reviewer workflow with full audit trail.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthResult:
      type: object
      properties:
        token: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            role: { type: string, enum: [admin, reviewer, user] }
    KycInput:
      type: object
      required: [fullName, dateOfBirth, nationality, documentType, documentNumber]
      properties:
        fullName: { type: string }
        dateOfBirth: { type: string, format: date }
        nationality: { type: string, minLength: 2, maxLength: 2 }
        documentType: { type: string, enum: [passport, id_card, driver_license] }
        documentNumber: { type: string }
    KycProfile:
      allOf:
        - $ref: '#/components/schemas/KycInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            userId: { type: string, format: uuid }
            status: { type: string, enum: [pending, in_review, approved, rejected] }
            riskScore: { type: integer }
            rejectionReason: { type: string, nullable: true }
paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: OK
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '409':
          description: Email already registered
          content:
            application/json: { schema: { $ref: '#/components/schemas/Error' } }
  /api/auth/login:
    post:
      summary: Exchange credentials for a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
  /api/kyc/profile:
    post:
      summary: Submit a KYC profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KycInput' }
      responses:
        '201':
          description: Profile created
          content:
            application/json: { schema: { $ref: '#/components/schemas/KycProfile' } }
  /api/kyc/profile/me:
    get:
      summary: Get the authenticated user's profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json: { schema: { $ref: '#/components/schemas/KycProfile' } }
  /api/kyc/review/queue:
    get:
      summary: List profiles awaiting manual review (reviewer/admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
  /api/kyc/review/{id}/decide:
    post:
      summary: Approve or reject a profile (reviewer/admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                reason: { type: string }
      responses:
        '200':
          description: Updated profile